search
The user is now able to search through the list of users in the system via email, if the user exists the user can be added (yet to be implemented). if the user doesn’t exist an error message will appear. pervious code was also changed for. a more simple and easy to understand and manipulatable approach. component didmount has been implemented at the home page. next to be completed, the user will be able to addd contacts and will be able to remove contacts