contacts completed
The previous code which was messy didn't allow for users to search for all contacts within the system. A new page was added which allows the user to search through the database, for example searching for a user by firstname, last name or email. the user can be added, removed or blocked. design is completed for the system up to this point, profile image is needed to be added for each user which can then be displayed next to the contacts