App is complete
All end points have now been met. The user is now able to send and receive messages in the correct format intended. Yet to be complete is uploading a profile picture and potentially a group chat picture. the code needs to be cleaned up and validation needs completing for the end points