Final Code
All comments have now been added within the code, the code now ensures when the user takes a picture and the back button is clicked the image updates on the profile page. Unnecessary code was also removed from the code, images were imported correctly rather than using the url of the image. my models were turned into functions rather than calling rendering a model and setting the states where i needed it, now i call the function which sets the states and opens the modal. finally the code checks if the user is logged in, if so the user is sent straight to the homepage whereas before the user was sent to the login page